+++ src/UI/Core/UIContainer.h
@@ -28,6 +28,8 @@ protected:
 	virtual void visibilityChanged(bool visible);
 	virtual void positionChanged(float dX, float dY);
 
+	virtual void onFocusChanged();
+
 private:
 	typedef std::map<std::string, Control*> ControlList;
 	//typedef std::vector<Control*> ControlList;