+++ src/UI/UI.h
@@ -6,3 +6,5 @@
 #include "Core/TextArea.h"
 #include "Core/TextField.h"
 #include "Core/UIContainer.h"
+#include "Core/Window.h"
+#include "Core/WindowStack.h"